Refactor: Quiz 화면 리펙토링
