+++ src/modules/quiz/types.d.ts
@@ -11,3 +11,5 @@ export type Quiz = {
 export type ItemStatus = 'correct' | 'wrong' | 'select' | 'idle';
 
 export type QuizStatus = 'default' | 'correct' | 'wrong';
+
+export type SwipeStatus = 'prev' | 'current' | 'next';
